+++ financial-dashboard-app/src/app/app.component.html
@@ -1,15 +1,4 @@
 <main class="main">
-  <div class="navigation"></div>
-  <div class="balance"></div>
-  <div class="date"></div>
-  <div class="account"></div>
-  <div class="total-net-worth"></div>
-  <div class="spendings"><app-spendings></app-spendings></div>
-  <div class="income"></div>
-  <div class="income-source"></div>
-  <div class="spending"><app-spending-overview></app-spending-overview></div>
-  <div class="income-goal"></div>
-  <div class="notifications-expenses-animals"></div>
-  <div class="income-expenses-graph"></div>
+  <app-grid-placement></app-grid-placement>
 </main>
 <router-outlet></router-outlet>
