+++ financial-dashboard-app/src/app/app.module.ts
@@ -3,14 +3,16 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { SpendingsComponent } from './spendings/spendings.component';
-import { SpendingOverviewComponent } from './spending-overview/spending-overview.component';
+import { SpendingsComponent } from './grid-placement/spendings/spendings.component';
+import { SpendingOverviewComponent } from './grid-placement/spending-overview/spending-overview.component';
+import { GridPlacementComponent } from './grid-placement/grid-placement.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     SpendingsComponent,
-    SpendingOverviewComponent
+    SpendingOverviewComponent,
+    GridPlacementComponent
   ],
   imports: [
     BrowserModule,
