+++ financial-dashboard-app/src/app/grid-placement/grid-placement.component.html
@@ -1,0 +1,15 @@
+<main class="main">
+  <div class="navigation"></div>
+  <div class="balance"></div>
+  <div class="date"></div>
+  <div class="account"></div>
+  <div class="total-net-worth"></div>
+  <div class="spendings"><app-spendings></app-spendings></div>
+  <div class="income"></div>
+  <div class="income-source"></div>
+  <div class="spending"><app-spending-overview></app-spending-overview></div>
+  <div class="income-goal"></div>
+  <div class="notifications-expenses-animals"></div>
+  <div class="income-expenses-graph"></div>
+</main>
+<router-outlet></router-outlet>
