+++ financial-dashboard-app/src/app/grid-placement/grid-placement.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-grid-placement',
+  templateUrl: './grid-placement.component.html',
+  styleUrl: './grid-placement.component.scss'
+})
+export class GridPlacementComponent {
+
+}
