+++ financial-dashboard-app/src/app/grid-placement/spending-overview/spending-overview.component.html
@@ -1,4 +1,4 @@
-<div class="spending-overview-container">
+<section class="spending-overview-container">
   <div class="spendings">
     <h3>Spendings</h3>
     <li class="spending-subjects">
@@ -40,4 +40,4 @@
       </div>
     </li>
   </div>
-</div>
+</section>
