+++ financial-dashboard-app/src/app/grid-placement/spending-overview/spending-overview.component.spec.ts
