+++ financial-dashboard-app/src/app/grid-placement/spendings/spendings.component.html
